disc-gym/requirements.txt
numpy~=1.19.2
termcolor~=1.1.0
pytorch~=1.7.0
gym~=0.17.3
tensorboard~=2.11.0